+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 Hydrophone is the module responsible for sending emails.
 This API sends notifications to users for things like forgotten passwords, initial signup, and invitations.
 
+## unreleased
+### changed
+- YLP-682 Add team information in received invitations
 ## 1.6.1 - 2021-05-04
 ### Changed
 - YLP-710 Modify route /accept/team/invite to acknowledge notifications that does not requires a specific action (i.e. change member role)